ensure dtype match between diffused latents and vae weights

+++ src/diffusers/pipelines/pixart_alpha/pipeline_pixart_sigma.py
@@ -862,7 +862,7 @@ class PixArtSigmaPipeline(DiffusionPipeline):
                         callback(step_idx, t, latents)
 
         if not output_type == "latent":
-            image = self.vae.decode(latents / self.vae.config.scaling_factor, return_dict=False)[0]
+            image = self.vae.decode(latents.to(self.vae.dtype) / self.vae.config.scaling_factor, return_dict=False)[0]
             if use_resolution_binning:
                 image = self.image_processor.resize_and_crop_tensor(image, orig_width, orig_height)
         else:
